requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
psycopg[binary]==3.3.2
